:art: use @scan operator
