+++ src/parse/traverse.c
@@ -28,7 +28,10 @@ ANN m_bool traverse_func_def(const Env env, const Func_Def def) {
 ANN m_bool traverse_union_def(const Env env, const Union_Def def) {
   if(!GET_FLAG(def, scan1))
     CHECK_BB(scan1_union_def(env, def))
+  if(!GET_FLAG(def, scan2))
   CHECK_BB(scan2_union_def(env, def))
+  if(!GET_FLAG(def, check))
+    CHECK_BB(check_union_def(env, def))
   return check_union_def(env, def);
 }
